splitting controllers into files and re-exporting files
