+++ controllers/contacts/listContact.js
@@ -1,0 +1,9 @@
+const { Contact } = require("../../schema");
+
+//Search all contacts in the database
+const listContacts = async (req, res) => {
+	const contacts = await Contact.find();
+	res.json(contacts);
+};
+
+module.exports = listContacts;
